# swaggerV2.json
{
  "swagger": "2.0",
  "info": {
    "description": "RESTfull API baseada em node.js Com o objetivo da partilha de viagens e experiencias vividas, classificações e comentáriosentre utilizadores. Disponível em https://tppweb.herokuapp.com",
    "version": "2.0.0",
    "title": "Trips Together",
    "contact": {
      "email": "o.nosso.email@gmail.com"
    }
  },
...
  "produces": [
    "application/json"
  ],
  "host": "tppweb.herokuapp.com",
  "basePath": "/api/v2",
  "paths": {
...
    "/momentos": {
      "post": {
        "tags": [
          "Momentos"
        ],
        "summary": "Adiciona uma novo momento",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "idViagem",
            "in": "formData",
            "description": "ID Viagem",
            "required": true,
            "type": "string"
          },
          {
            "name": "pais",
            "in": "formData",
            "description": "Pais de destino",
            "type": "string"
          },
          {
            "name": "cidade",
            "in": "formData",
            "description": "Cidade de destino",
            "type": "string"
          },
          {
            "name": "latitude",
            "in": "formData",
            "description": "Coordenadas de latitude",
            "type": "string"
          },
          {
            "name": "longitude",
            "in": "formData",
            "description": "Coordenadas de longitude",
            "type": "string"
          },
          {
            "name": "dataInicio",
            "in": "formData",
            "description": "Data de inicio do momento",
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "dataFim",
            "in": "formData",
            "description": "Data de fim do momento",
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "descricao",
            "in": "formData",
            "description": "Descrição do momento",
            "type": "string"
          },
          {
            "name": "narrativa",
            "in": "formData",
            "description": "Narrativa do momento",
            "type": "string"
          },
          {
            "name": "ficheiros",
            "in": "formData",
            "description": "imagem associada ao momento",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso ao gravar dados",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
...
    "/momentos/{id}": {
      "get": {
        "tags": [
          "Momentos"
        ],
        "summary": "Obter um momento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID momento",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Momentos"
        ],
        "summary": "Atualiza um momento existente",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID momento",
            "required": true,
            "type": "string"
          },
          {
            "name": "pais",
            "in": "formData",
            "description": "Pais de destino",
            "type": "string"
          },
          {
            "name": "cidade",
            "in": "formData",
            "description": "Cidade de destino",
            "type": "string"
          },
          {
            "name": "latitude",
            "in": "formData",
            "description": "Coordenadas de latitude",
            "type": "string"
          },
          {
            "name": "longitude",
            "in": "formData",
            "description": "Coordenadas de longitude",
            "type": "string"
          },
          {
            "name": "dataInicio",
            "in": "formData",
            "description": "Data de inicio do momento",
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "dataFim",
            "in": "formData",
            "description": "Data de fim do momento",
            "type": "string",
            "format": "date-time"
          },
          {
            "name": "descricao",
            "in": "formData",
            "description": "Descrição do momento",
            "type": "string"
          },
          {
            "name": "narrativa",
            "in": "formData",
            "description": "Narrativa do momento",
            "type": "string"
          },
          {
            "name": "ficheiros",
            "in": "formData",
            "description": "imagem associada ao momento",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso ao gravar dados",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Momentos"
        ],
        "summary": "Remove um momento por id",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID momento",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Momento removida com sucesso",
            "schema": {
              "$ref": "#/definitions/Response"
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
...
    "/classificacao/momento/{id}": {
      "get": {
        "tags": [
          "Classificação"
        ],
        "summary": "Obter classificações de um momento",
        "description": "Obtem todos as classificações a um determinado momento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do momento",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
...
    "/comentarios/momento/{id}": {
      "get": {
        "tags": [
          "Comentários"
        ],
        "summary": "Obter os comentários de um momento",
        "description": "Obtem todos os comentários de um determinado momento",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do momento",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Resposta com sucesso",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "400": {
            "description": "Resposta inválida",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          },
          "401": {
            "description": "Utilizador não autenticado",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Response"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Response": {
      "type": "object",
      "properties": {
        "ErroCorde": {
          "type": "string"
        },
        "Message": {
          "type": "string"
        },
        "Error": {
          "type": "object"
        },
        "Data": {
          "type": "object"
        }
      }
    }
  }
}